// src/transactions/enums/transaction.enums.ts
export enum Categories {
  IMPUESTOS = 'IMPUESTOS',
  COMIDA = 'COMIDA',
  TRANSPORTE = 'TRASPORTE',
  VIVIENDA = 'VIVIENDA',
  ESTUDIOS = 'ESTUDIOS',
  DEPORTES = 'DEPORTES',
  OCIO = 'OCIO',
  SERVICIO = 'SERVICIO',
  VARIOS = 'VARIOS',
}

export enum MeansOfPayment {
  EFECTIVO = 'EFECTIVO',
  DEVITO = 'DEVITO',
  CREDITO = 'CREDITO',
  TRANSFERENCIA = 'TRANSFERENCIA',
  CHEQUE = 'CHEQUE',
  OTROS = 'OTROS',
}

export enum Area {
  ADMINISTRACION = 'ADMINISTRACIÓN',
  COMPRAS = 'COMPRAS',
  VENTAS = 'VENTAS',
  OTROS = 'OTROS',
}